feat: add face extraction mode to view_align operator and rename it to face_sketch, rename new_sketch to blank_sketch, and sketch_bevel to vertex_bevel

+++ face_sketch.py
@@ -1,23 +1,30 @@
 import bpy
 import bmesh
-from . overlay import update_overlay, init_overlay, register_draw_handler, unregister_draw_handler, draw_header, draw_hint
+from math import radians
+from . overlay import update_overlay, init_overlay, register_draw_handler, unregister_draw_handler, draw_header, draw_hint, draw_property
 from . utils import averaged_vector, set_3d_cursor, create_rotation_matrix_from_face
 
 
-class ND_OT_view_align(bpy.types.Operator):
-    bl_idname = "nd.view_align"
-    bl_label = "View Align"
-    bl_description = "Orientates the view so it is looking directly at a face, sets the 3D cursor, and adds a custom transform orientation."
+class ND_OT_face_sketch(bpy.types.Operator):
+    bl_idname = "nd.face_sketch"
+    bl_label = "Face Sketch"
+    bl_description = "Orientate the view to a face, and optionally lift a copy of the face as a new starting sketch"
     bl_options = {'UNDO'}
 
 
     def modal(self, context, event):
-        if not self.pin_overlay and event.type == 'MOUSEMOVE':
-            update_overlay(self, context, event)
-        
-        elif event.type == 'P' and event.value == 'PRESS':
+        self.key_alt = event.alt
+
+        if event.type == 'P' and event.value == 'PRESS':
             self.pin_overlay = not self.pin_overlay
-            update_overlay(self, context, event, pinned=self.pin_overlay, x_offset=270, lines=1)
+
+        elif event.type == 'WHEELUPMOUSE':
+            if event.alt:
+                self.face_operation_mode = (self.face_operation_mode + 1) % 2
+            
+        elif event.type == 'WHEELDOWNMOUSE':
+            if event.alt:
+                self.face_operation_mode = (self.face_operation_mode - 1) % 2
 
         elif event.type == 'LEFTMOUSE':
             return {'PASS_THROUGH'}
@@ -28,17 +35,23 @@ class ND_OT_view_align(bpy.types.Operator):
             return {'FINISHED'}
 
         elif event.type in {'RIGHTMOUSE', 'ESC'}:
-            self.revert(context)
+            self.clean_up(context)
 
             return {'CANCELLED'}
         
         elif event.type == 'MIDDLEMOUSE' or (event.alt and event.type in {'LEFTMOUSE', 'RIGHTMOUSE'}) or event.type.startswith('NDOF'):
             return {'PASS_THROUGH'}
 
+        update_overlay(self, context, event, pinned=self.pin_overlay, x_offset=320, lines=2)
+
         return {'RUNNING_MODAL'}
 
 
     def invoke(self, context, event):
+        self.face_operation_mode = 0 # Sketch Only (0), Extract & Sketch (1)
+
+        self.key_alt = False
+
         self.prepare_face_selection_mode(context)
 
         self.pin_overlay = False
@@ -72,6 +85,9 @@ class ND_OT_view_align(bpy.types.Operator):
 
         bpy.ops.object.mode_set_with_submode(mode='EDIT', mesh_select_mode={'FACE'})
 
+        context.object.name = 'ND — Face Extract'
+        context.object.data.name = 'ND — Face Extract'
+
     
     def set_3d_cursor_to_face(self, context):
         mesh = bmesh.from_edit_mesh(context.object.data)
@@ -90,19 +106,36 @@ class ND_OT_view_align(bpy.types.Operator):
 
 
     def set_face_transform_orientation(self):
-        bpy.ops.transform.create_orientation(name="ND — " + ND_OT_view_align.bl_label, use=True)
+        bpy.ops.transform.create_orientation(name="ND — Sketch Surface", use=True)
 
 
-    def finish(self, context):
+    def prepare_face_sketch(self, context):
         bpy.ops.view3d.view_axis(type='TOP', align_active=True)
         self.set_face_transform_orientation()
         self.set_3d_cursor_to_face(context)
-        self.revert(context)
+
+
+    def isolate_face(self, context):
+        bpy.ops.mesh.select_all(action='INVERT')
+        bpy.ops.mesh.delete(type='FACE')
+        bpy.ops.mesh.select_all(action='SELECT')
+
+
+    def finish(self, context):
+        self.isolate_face(context)
+        self.prepare_face_sketch(context)
+        self.clean_up(context)
 
     
-    def revert(self, context):
+    def clean_up(self, context):
         bpy.ops.object.mode_set(mode='OBJECT')
-        bpy.ops.object.delete()
+        bpy.ops.object.origin_set(type='ORIGIN_GEOMETRY', center='MEDIAN')
+
+        context.object.show_in_front = False
+
+        if self.face_operation_mode == 0:
+            bpy.ops.object.delete()
+
         unregister_draw_handler(self)
     
 
@@ -111,18 +144,25 @@ def draw_text_callback(self):
 
     draw_hint(self, "Select a single face...", "Press space to confirm")
 
+    draw_property(
+        self,
+        "Face Mode: {0}".format(['Sketch Only', 'Extract & Sketch'][self.face_operation_mode]),
+        "Alt (Sketch Only, Extract & Sketch)",
+        active=(self.key_alt),
+        alt_mode=False)
+
 
 def menu_func(self, context):
-    self.layout.operator(ND_OT_view_align.bl_idname, text=ND_OT_view_align.bl_label)
+    self.layout.operator(ND_OT_face_sketch.bl_idname, text=ND_OT_face_sketch.bl_label)
 
 
 def register():
-    bpy.utils.register_class(ND_OT_view_align)
+    bpy.utils.register_class(ND_OT_face_sketch)
     bpy.types.VIEW3D_MT_object.append(menu_func)
 
 
 def unregister():
-    bpy.utils.unregister_class(ND_OT_view_align)
+    bpy.utils.unregister_class(ND_OT_face_sketch)
     bpy.types.VIEW3D_MT_object.remove(menu_func)
     unregister_draw_handler(self)
